api/pyproject.toml
[tool.poetry]
name = "api"
version = "0.1.0"
description = ""
package-mode = false
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
annotated-types = "0.7.0"
anyio = "4.4.0"
asgiref = "3.8.1"
certifi = "2024.8.30"
colorama = "0.4.6"
distro = "1.9.0"
dj-rest-auth = "6.0.0"
django = "5.1"
django-cors-headers = "4.4.0"
django-environ = "0.11.2"
django-storages = {extras = ["s3"], version = "^1.14.4"}
djangorestframework = "3.15.2"
h11 = "0.14.0"
httpcore = "1.0.5"
httpx = "0.27.2"
idna = "3.10"
jiter = "0.5.0"
jmespath = "1.0.1"
openai = "1.46.0"
pillow = "10.4.0"
pydantic = "2.9.2"
pydantic-core = "2.23.4"
python-dateutil = "2.9.0.post0"
six = "1.16.0"
sniffio = "1.3.1"
sqlparse = "0.5.1"
tqdm = "4.66.5"
typing-extensions = "4.12.2"
tzdata = "2024.1"
urllib3 = "2.2.2"
celery = "^5.4.0"
beautifulsoup4 = "^4.12.3"
django-redis = "^5.4.0"
gunicorn = "^23.0.0"
s3transfer = "0.10.2"
boto3 = "1.35.14"
botocore = "1.35.14"
rarfile = "^4.2"
django-allauth = {extras = ["socialaccount"], version = "^65.2.0"}
python-jose = "^3.3.0"
google-api-python-client = "^2.154.0"
black = "^24.10.0"
isort = "^5.13.2"
bandit = "^1.8.0"
flake8 = "^7.1.1"

[tool.poetry.group.dev]
optional = true

[tool.poetry.group.dev.dependencies]
psycopg2 = "2.9.9"

[tool.poetry.group.prod]
optional = true

[tool.poetry.group.prod.dependencies]
psycopg2-binary = "2.9.9"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"
line_length = 79
